Add stub of Treasury contract
This commit adds Debt and Surplus auctions endpoints and validators.
Both are not finished yet but it gives some overview of how they would
work. There's a lot of mock-up currency flowing around (i.e. DANA
shouldn't be minted at the start of the Treasury contract).

The PAB simulator shows a demo of debt and surplus auctions.