utxo-design: Add LiquidationAct
This doesn't use the same algorithm as MakerDAO, as that would require on-chain auctions.
On-chain auctions require a lot of transactions and UTXOs, and a lot of logic to support
them.

Considering the state of Cardano and the limited time available, this alternative approach
seems more optimal.

In essence, there is a fixed reward rate for liquidators, which makes it most rewardable
to liquidate as soon as possible. Since there is no auction, there is no system to select
who gets to liquidate a vault.

Please do some proper review of the algorithm, as I haven't formalised this in any proof
assistant.

CC @quinn-dougherty ?